Cleaned up parser and moved comments
